+++ R/read_credentials.R
@@ -9,7 +9,7 @@
 #' @author Thomas J. Leeper <thosjleeper@gmail.com>
 #' @references
 #'   \href{https://blogs.aws.amazon.com/security/post/Tx3D6U6WSFGOK2H/A-New-and-Standardized-Way-to-Manage-Credentials-in-the-AWS-SDKs}{Amazon blog post describing the format}
-#' @seealso \code{\link{signature_v2_auth}}, \code{\link{use_credentials}}
+#' @seealso \code{\link{signature_v2_auth}}
 #' @examples
 #' \dontrun{
 #' # set environment variables from a profile
@@ -51,6 +51,8 @@ use_credentials <- function(profile = "default", file = default_credentials_file
     invisible(x)
 }
 
+#' @rdname read_credentials
+#' @export
 default_credentials_file <- function() {
     if (.Platform[["OS.type"]] == "windows") {
         home <- Sys.getenv("USERPROFILE")