+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES TO aws.signature 0.2.8
+
+* Added a `locate_credentials()` function to walk through a hierarchy of possible credential locations, beginning with user-supplied values, then environment variables, IAM roles, local or global credentials files, and finally (if applicable) an EC2 role for the currently running instance. (#11)
+
 # CHANGES TO aws.signature 0.2.7
 
 * `read_credentials()` now looks for the credentials file in a more reasonable location on Windows (#12/#13, h/t user:kadrach)